+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## 0.6.2 - 12. September 2020
+### Fixed
+- `[builder]` Fixed issue with context files being absolute.
+
+## 0.6.2 - 12. September 2020
 ### Added
 - `[builder]` now compiles into the relative directory from its source file.